requirements.txt
streamlit
requests
beautifulsoup4
pypdf
openai
anthropic
google-generativeai

